% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HMM.R
\name{compute_ML_rho}
\alias{compute_ML_rho}
\title{Maximum likelihood of recombination rate}
\usage{
compute_ML_rho(Y, M, theta = NULL, rho0 = 1e-08)
}
\arguments{
\item{Y}{A binary sequence classifying all observed sites as either
homozygous (encoded Y[i] == 0) or heterozygous (encoded Y[i] == 1)}

\item{M}{the number of hidden states to be used in the HMM}

\item{theta}{The mutation rate to be used (defaults to Watterson's
estimator).}

\item{rho0}{initial guess used as starting point of gradiant-algorithm.}
}
\value{
a maximum likelihood estimate of the recombination rate between two
  sites.
}
\description{
Approximate the maximum likelihood estimate of the recombination rate.
This is done using nlm() for newtonian maximization of the likelihood
function.
}
