% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HMM.R
\name{compute_mu}
\alias{compute_mu}
\title{Compute emission probabilities for HMM}
\usage{
compute_mu(M, theta)
}
\arguments{
\item{M}{the number of states of the hidden variable X}

\item{theta}{the rate of mutation between two lineages}
}
\value{
A vector mu s.t. mu[i] = P(Y = 1 | X = i) for i=1...M
}
\description{
Compute emission probabilities for HMM
}
