% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HMM.R
\name{compute_Px}
\alias{compute_Px}
\title{Compute hidden state transition kernel}
\usage{
compute_Px(M, rho)
}
\arguments{
\item{M}{the number of hissden states}

\item{rho}{The recombination rate between two sites}
}
\value{
An M-by-M matrix P, s.t. \code{P[i,j]} = P(X_t = j | X_{t-1} = i)
}
\description{
A wrapper function for calling A. Hobolths code for computin the hiden state
rate matrix by way of matrix exponentiation (as outlined in his note).
}
