% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HMM.R
\name{count_changepoints}
\alias{count_changepoints}
\title{Count changepoints}
\usage{
count_changepoints(vec)
}
\arguments{
\item{vec}{A vector to be scanned}
}
\value{
a positive integer
}
\description{
Scan along a sequence and count the number of times an entry does not match
the previous entry
}
