% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HMM.R
\name{count_runs}
\alias{count_runs}
\title{Count runs}
\usage{
count_runs(vec)
}
\arguments{
\item{vec}{A vector to be scanned}
}
\value{
The number of runs
}
\description{
Scan along a sequence; return the number of intervals wherein ajacent entries
are identical
}
