% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HMM.R
\name{d_seq}
\alias{d_seq}
\title{Generate discretization timepoints}
\usage{
d_seq(M)
}
\arguments{
\item{M}{The number of intervals in the discretization; should be a positive
integer}
}
\value{
A vector of upper limits of each interval. last element will be
\code{Inf}
}
\description{
Compute the sequence of discretization timepoints (cf. Asger's note)
}
