% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HMM.R
\name{forward_algorithm}
\alias{forward_algorithm}
\title{Forward algorithm}
\usage{
forward_algorithm(Y, transition_probs, emission_probs, prior)
}
\arguments{
\item{Y}{sequence of observed emissions}

\item{emission_probs}{a vector of emission probabilities. The ith entry is
interpreted as the probability of Y=1 given X=i.}

\item{prior}{a vector encoding the a prior distribution on X0; not X[1], but
the stete preceding it, i.e. there is no observed emission of the variable X0}

\item{trnasition_probs}{an M-by-M matrix of transition probabilities}
}
\value{
a vector alpha with entries P(X_T = i, Y_1:T = y_1:T) for i=1...M.
}
\description{
Computes alpha_T[i] = P(X_T = i, Y_1:T = y_1:T) for i=1...M, where M is the
number of hidden states, using the so-called forward-algorithm. This
implementation is less numerically stable than
\code{forward_algorithm_logspace}, which is otherwise identical, but works in
logspace instead.
}
