% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HMM.R
\name{log_likelihood_Y}
\alias{log_likelihood_Y}
\title{Marginal log-likelihood of Y}
\usage{
log_likelihood_Y(Y, M, rho = NULL, theta = NULL, Px = NULL, mu = NULL,
  prior_x0 = NULL)
}
\arguments{
\item{Y}{observed sequence of emissions (1: heterozygous site;
0: homozygous site).}

\item{M}{size of hidden statespace; an integer.}

\item{rho}{the recombinartion rate between two ajacent sites.}

\item{theta}{mutation rate between two sequences at any given position.}

\item{Px}{transition kernel of hidden variables. May be passed instead of
rho.}

\item{mu}{a vector of emission probabilities. The ith entry is
interpreted as the probability of Y=1 given X=i. May be passed instead of
theta}

\item{x0_prior}{initial distribution of the hidden markov chain. Defaults to
a uniform distribution.}
}
\value{
a single number, corresponding to l(theta,rho|Y)
}
\description{
Marginalizes out the hidden variables X using the forward algorithm to
compute the log-likelihood of an observed sequence of emissions.
}
