% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HMM.R
\name{log_sum_exp}
\alias{log_sum_exp}
\title{Compute log-sum-exp}
\usage{
log_sum_exp(x)
}
\arguments{
\item{x}{a vector of numbers}
}
\value{
log(exp(x[1]) + exp(x[2]) + ... + exp(x[n]))
}
\description{
Compute log(exp(x1) + exp(x2) + ... + exp(xn)) in a numerically stable
manner by relying on the identity
log(exp(x1) + exp(x2) + ... + exp(xn))
= c + log(exp(x1 - c) + exp(x2 - c) + ... + exp(xn - c))
for any real c.
We here work with c = (min(x) + max(x)) / 2, which minimizes
max_i |x_i - c|
}
