% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HMM.R
\name{log_sum_exp_lower_bound}
\alias{log_sum_exp_lower_bound}
\title{Approximate log-sum-exp from below}
\usage{
log_sum_exp_lower_bound(x)
}
\arguments{
\item{x}{a vector of numbers}
}
\value{
max(x)
}
\description{
we rely on the approximation
max(x)
<=
log(exp(x1) + exp(x2) + ... + exp(xn))
<=
max(x)+log(n)
}
