% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HMM.R
\name{sample_X}
\alias{sample_X}
\title{Sample hidden states}
\usage{
sample_X(sample_X0, L, Px)
}
\arguments{
\item{sample_X0}{A prior distribution on X0; sample_X0() should return an
integer in the range 1:(dim(Px)[1]); e.g. “function() sample(M,1)” for a
uniform distribution on 1:M or “function() i” for a point mass at i.}

\item{L}{the length of the desired sequence of samples}

\item{Px}{The transition rates of the markov chain X to be sampled from.}
}
\value{
A vector of samples from X of length \code{L}.
}
\description{
Sample X consequtive states from the hidden markov model
}
