% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HMM.R
\name{sample_YgivenX}
\alias{sample_YgivenX}
\title{Sample emissions}
\usage{
sample_YgivenX(X, theta, M)
}
\arguments{
\item{X}{a vector of integers encoding the states of the hidden variables}

\item{theta}{the rate of mutation between two sequences}

\item{M}{the number of states of the hidden variables X}
}
\value{
A vector of the same length as \code{X} s.t. Y[i] = 1 with
  probability \code{mu[X[i]]}.
}
\description{
Samples a sequence of binary emissions (Y[i] = 1 is interpreted as,
“Heterozygous sites at position i”; whereas Y[i] = 0 is interpreted as
signifying homozugosity.)
}
