% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HMM.R
\name{theta_watterson}
\alias{theta_watterson}
\title{Watterson's estimator}
\usage{
theta_watterson(s, n, L)
}
\arguments{
\item{s}{number of segregating sites observed.}

\item{n}{number of sequences sampled.}

\item{L}{number of sites (segregating + non-segregating) observed in sample.}
}
\value{
A non-negative floating point number
}
\description{
Compute Watterson's estimator of the site-wise mutation rate
}
