% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HMM.R
\name{compute_ML_rho}
\alias{compute_ML_rho}
\title{Maximum likelihood of recombination rate}
\usage{
compute_ML_rho(Y, M, theta = NULL, magnitude_max = 1)
}
\arguments{
\item{Y}{A binary sequence classifying all observed sites as either
homozygous (encoded Y[i] == 0) or heterozygous (encoded Y[i] == 1)}

\item{M}{the number of hidden states to be used in the HMM}

\item{theta}{The mutation rate to be used (defaults to Watterson's
estimator).}

\item{magnitude_max}{upper magnitude bound on rho, i.e. magnitude_max = k
indicates that we firmly believe rho<10^k to be the case; defaults to +1.}
}
\value{
a maximum likelihood estimate of the recombination rate between two
sites.
}
\description{
Approximate the maximum likelihood estimate of the recombination rate.
This is done in two steps (each of which relies on optimize()): one to find
the approximate scale of rho; followed by a step to find the maximum
likelihood on that scale.
}
